Working on the conversion of WLO data timestamps from 5mins intervalls to be compatible with 3mins time intervalls.

+++ ca/gc/dfo/chs/wltools/util/MeasurementCustom.java
@@ -84,13 +84,13 @@ final public class MeasurementCustom {
   // --- IMPORTANT: Do not use this method for obs MeasurementCustom data
   //     since we could have missing data so the result of this method could
   //     be dangerously misleading this kind of data.
-  public final static long getDataTimeIntervallSeconds(final List<MeasurementCustom> mcList) {
+  public final static int getDataTimeIntervallSeconds(final List<MeasurementCustom> mcList) {
 
     final long firstTimeStampSeconds= mcList.get(0).getEventDate().getEpochSecond();
     final long secondTimeStampSeconds= mcList.get(1).getEventDate().getEpochSecond();
 
     // --- We do not assume that the secondTimeStampSeconds value is larger than the
     //     firstTimeStampSeconds value so usr Math.abs here.
-    return Math.abs(secondTimeStampSeconds - firstTimeStampSeconds);
+    return (int) Math.abs(secondTimeStampSeconds - firstTimeStampSeconds);
   }
 }
